requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
PyJWT
cryptography
requests